backend/requirements.txt
Django>=5.1,<6

djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.4,<5
django-environ>=0.11,<1
dj-database-url>=2.2,<3
psycopg[binary]>=3.2,<4
PyJWT>=2.9,<3
requests>=2.32,<3
groq>=0.11,<1
whitenoise>=6.7,<7
gunicorn>=22,<23
pypdf>=5.1,<6
python-docx>=1.1,<2
llama-index>=0.11,<0.12
llama-index-llms-groq>=0.2,<0.3
llama-index-embeddings-openai>=0.2,<0.3
tavily-python>=0.5,<0.6
pytest>=8.2,<9
pytest-django>=4.9,<5
pytest-cov>=5.0,<6
factory-boy>=3.3,<4
Faker>=26,<30
